Daniel_pallares/requirements.txt

Django==3.2.6
psycopg2-binary==2.8
djangorestframework==3.12.2